quizlet-main/backend/requirements.txt
django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
numpy==1.26.3
packaging==24.0
pandas==2.2.0
charset-normalizer==3.3.2
python-dateutil==2.8.2
psycopg2-binary==2.9.9
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.3.1
oauthlib==3.2.2
requests-toolbelt==1.0.0
sqlparse==0.4.4
urllib3==2.2.1
openpyxl==3.1.2
drf-yasg==1.21.5
setuptools==72.1.0